refactor: remove reducers from protocol base class
