+++ src/vector/backends/awkward.py
@@ -1240,6 +1240,14 @@ class VectorArray2D(VectorAwkward2D, ak.Array):  # type: ignore[misc]
         """Like ``np.ndarray.allclose``, but for VectorArray2D."""
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
 
+    def sum(self) -> VectorProtocol:
+        from vector._compute.planar import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.rho2, axis=-1)
+
 
 behavior["*", "Vector2D"] = VectorArray2D
 
@@ -1273,6 +1281,14 @@ class VectorArray3D(VectorAwkward3D, ak.Array):  # type: ignore[misc]
     ) -> BoolCollection:
         """Like ``np.ndarray.allclose``, but for VectorArray3D."""
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
+
+    def sum(self) -> VectorProtocol:
+        from vector._compute.spatial import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.mag2, axis=-1)
 
 
 behavior["*", "Vector3D"] = VectorArray3D
@@ -1308,6 +1324,14 @@ class VectorArray4D(VectorAwkward4D, ak.Array):  # type: ignore[misc]
         """Like ``np.ndarray.allclose``, but for VectorArray4D."""
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
 
+    def sum(self) -> VectorProtocol:
+        from vector._compute.lorentz import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.tau2, axis=-1)
+
 
 behavior["*", "Vector4D"] = VectorArray4D
 
@@ -1342,6 +1366,14 @@ class MomentumArray2D(MomentumAwkward2D, ak.Array):  # type: ignore[misc]
         """Like ``np.ndarray.allclose``, but for MomentumArray4D."""
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
 
+    def sum(self) -> VectorProtocol:
+        from vector._compute.planar import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.rho2, axis=-1)
+
 
 behavior["*", "Momentum2D"] = MomentumArray2D
 
@@ -1375,6 +1407,14 @@ class MomentumArray3D(MomentumAwkward3D, ak.Array):  # type: ignore[misc]
     ) -> BoolCollection:
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
 
+    def sum(self) -> VectorProtocol:
+        from vector._compute.spatial import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.mag2, axis=-1)
+
 
 behavior["*", "Momentum3D"] = MomentumArray3D
 
@@ -1407,6 +1447,14 @@ class MomentumArray4D(MomentumAwkward4D, ak.Array):  # type: ignore[misc]
         equal_nan: BoolCollection = False,
     ) -> BoolCollection:
         return ak.all(self.isclose(other, rtol=rtol, atol=atol, equal_nan=equal_nan))
+
+    def sum(self) -> VectorProtocol:
+        from vector._compute.lorentz import sum
+
+        return sum.dispatch(self)
+
+    def count_nonzero(self) -> ScalarCollection:
+        return self.lib.count_nonzero(self.tau2, axis=-1)
 
 
 behavior["*", "Momentum4D"] = MomentumArray4D
